added tests and refactored for reusability
